+++ auth.go
@@ -277,7 +277,11 @@ func isExpectedOAuthUserResponseURL(response *http.Response) bool {
 }
 
 func isOAuthUserJSONResponse(response *http.Response) bool {
-	mediaType, _, err := mime.ParseMediaType(response.Header.Get("Content-Type"))
+	contentTypes := response.Header.Values("Content-Type")
+	if len(contentTypes) != 1 {
+		return false
+	}
+	mediaType, _, err := mime.ParseMediaType(contentTypes[0])
 	if err != nil {
 		return false
 	}